Remove cites to non-app layer secure transports.
There have been a lot of comments about the inclusion of
WireGuard and Tor as potential secure transports, but I
actually think it's confusing to cite IPsec as well
for layering issues. We don't have a real praxis for
using things that aren't application-layer style protocols
for protecting this kind of traffic, so I don't think
we should encourage these. I've trimmed it to just
QUIC and TLS.