Remove essentially all uses of "tree".
In Eric Vyncke's DISCUSS he correctly notes that LeafNode is used
without being defined. I fixed that by just removing the term
but more generally the problem is the use of "tree" without
it being defined. Obviously the tree is an important concept
but that's really a matter for mls-protocol not mls-architecture
which is just concerned with group state as an abstract structure.

This change removes essentially all uses of the term tree except
for:

- In the final security considerations section where I think it's
  OK in context.

- In this sentence

  "A maximum number of steps that clients will move a secret tree ratchet
  forward in response to a single message before rejecting it."

I'm pretty sure that this can be reworded but I wasn't quite sure
the best way.