Disclaim replay protection
As [discussed on the MLS mailing list](https://mailarchive.ietf.org/arch/msg/mls/SRAV33J5Ka9A4R-Ymv8a_T_TXhE/), an attacker with access to the groups symmetric secrets can create a new PrivateMessage containing the content of a previously sent message, and a different generation value.  This PR clarifies that this sort of replay is not protected against by the protocol.